bug fixes transactions

+++ src/common/blockchain/interface-blockchain/blocks/Interface-Blockchain-Block-Data-Transactions.js
@@ -3,6 +3,7 @@ import WebDollarCrypto from 'common/crypto/WebDollar-Crypto'
 import consts from 'consts/const_global'
 import Serialization from 'common/utils/Serialization'
 import InterfaceBlockchainTransaction from "../transactions/transaction/Interface-Blockchain-Transaction";
+import Blockchain from "main-blockchain/Blockchain";
 
 class InterfaceBlockchainBlockDataTransactions {
 
@@ -20,8 +21,14 @@ class InterfaceBlockchainBlockDataTransactions {
 
     destroyBlockDataTransactions(){
 
-        for (let i=0; i<this.transactions.length; i++)
-            this.transactions[i].destroyTransaction();
+        for (let i=0; i<this.transactions.length; i++) {
+
+            if ( !Blockchain.blockchain.transactions.pendingQueue.findPendingTransaction(this.transactions[i]) )
+                this.transactions[i].destroyTransaction();
+
+            this.transactions[i] = undefined;
+
+        }
 
     }
